fix: removing custom emojie reactions

+++ internals/backend/controllers/rules_controller.go
@@ -75,7 +75,7 @@ func (rc *RulesController) getReactions(w http.ResponseWriter, r *http.Request) 
 }
 
 func (rc *RulesController) postReactions(w http.ResponseWriter, r *http.Request) {
-	rRules, ok := r.Context().Value(middleware.ValidateJsonCtxKey).([]rule.ReactionRule)
+	rRules, ok := middleware.JsonFromContext(r.Context()).([]rule.ReactionRule)
 
 	if !ok {
 		rc.logger.Error(common.ErrInternal, map[string]any{"details": "error while validating postReactions"})
@@ -128,7 +128,13 @@ func (rc *RulesController) postReactions(w http.ResponseWriter, r *http.Request)
 
 func (rc *RulesController) deleteReactions(w http.ResponseWriter, r *http.Request) {
 	gId := chi.URLParam(r, "id")
-	query := r.Context().Value(middleware.ValidateQueryCtxKey).([]rule.DeleteReactionRuleQuery)
+	query, ok := middleware.QueryFromContext(r.Context()).([]rule.DeleteReactionRuleQuery)
+
+	if !ok {
+		rc.logger.Error(common.ErrInternal, map[string]any{"details": "no value found in context"})
+		common.SendInternalError(w)
+		return
+	}
 
 	err := rc.reactionService.DeleteReactionRules(query, gId)
